+++ src/utils/Pyramid.cpp
@@ -66,7 +66,7 @@ ComparatorLevel compLevelAsc =
     };
 
 
-bool Pyramid::parse(json11::Json& doc, std::map<std::string, TileMatrixSet*> tmsList) {
+bool Pyramid::parse(json11::Json& doc) {
 
     // TMS
     std::string tmsName;
@@ -77,12 +77,11 @@ bool Pyramid::parse(json11::Json& doc, std::map<std::string, TileMatrixSet*> tms
         return false;
     }
 
-    std::map<std::string, TileMatrixSet*>::iterator tmsIt= tmsList.find ( tmsName );
-    if ( tmsIt == tmsList.end() ) {
-        errorMessage =  "Pyramid use unknown TMS [" + tmsName + "]" ;
+    tms = TmsBook::get_tms(tmsName);
+    if ( tms == NULL ) {
+        errorMessage =  "Pyramid use unknown or unloadable TMS [" + tmsName + "]" ;
         return false;
     }
-    tms = tmsIt->second;
 
     // FORMAT
     std::string formatStr;
@@ -205,14 +204,24 @@ bool Pyramid::parse(json11::Json& doc, std::map<std::string, TileMatrixSet*> tms
     return true;
 }
 
-Pyramid::Pyramid(Context* context, std::string path, std::map<std::string, TileMatrixSet*> tmsList) : Configuration(path) {
+Pyramid::Pyramid(std::string path) : Configuration(path) {
 
     nodataValue = NULL;
 
     /********************** Read */
 
+    ContextType::eContextType storage_type;
+    std::string tray_name, fo_name;
+    ContextType::split_path(path, storage_type, fo_name, tray_name);
+
+    Context* context = StoragePool::get_context(storage_type, tray_name);
+    if (context == NULL) {
+        errorMessage = "Cannot add " + ContextType::toString(storage_type) + " storage context to read pyramid's descriptor";
+        return;
+    }
+
     int size = -1;
-    uint8_t* data = context->readFull(size, filePath);
+    uint8_t* data = context->readFull(size, fo_name);
 
     if (size < 0) {
         errorMessage = "Cannot read descriptor "  + path ;
@@ -220,19 +229,17 @@ Pyramid::Pyramid(Context* context, std::string path, std::map<std::string, TileM
         return;
     }
 
-    /********************** Parse */
-
     std::string err;
     json11::Json doc = json11::Json::parse ( std::string((char*) data, size), err );
     if ( doc.is_null() ) {
-        errorMessage = "Cannot load JSON file "  + filePath + " : " + err ;
+        errorMessage = "Cannot load JSON file "  + path + " : " + err ;
         return;
     }
     if (data != NULL) delete[] data;
 
     /********************** Parse */
 
-    if (! parse(doc, tmsList)) {
+    if (! parse(doc)) {
         return;
     }